backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
transformers==4.30.2
requests==2.28.1 